mobile optimisation
On Desktop the width of the columns is still defined in percent, but on mobile the browser determines the needed width.
On mobile buttons stack on top of each other and not besides.
On mobile the id column , the bytes uploaded so far and the progressbar get hidden.
On both mobile and desktop file names get shortened (like "foobar...") if theres not enough space available.
On mobile the user can not zoom anymore (makes it feel less like a website but more like a app).